Implementation/requirements.txt
singleton-decorator
psycopg2-binary